+++ src/main/java/com/snsapi/user/UserServiceInterface.java
@@ -8,7 +8,8 @@ import java.util.Optional;
 public interface UserServiceInterface {
     List<User> findAll();
 
-    User findById(Long id);
+    User findById(Integer id);
+
 
     User save(User user);
 
@@ -17,4 +18,6 @@ public interface UserServiceInterface {
     UserDetails loadUserByUsername(String username);
 
     Optional<User> findByUserEmail(String username);
+
 }
+
